The status code is a string and cannot be compared with '<'
The variable "status_code" is a string that contains a textual
description of the status, e.g. '404 NOT FOUND' or '500 INTERNAL
SERVER ERROR'. It is therefore better to use the variable 'code' that
only contains the status code, e.g. '404' or '500'. Plus, both
variables are actually strings, so in the rare cases where
'content-length' is not set as a header, the code reaches this
comparison and will crash, outputting an error in the console but not
triggering the in-browser debugger.